backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.31
psycopg2-binary==2.9.10
alembic==1.13.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
email-validator==2.3.0
